+++ protocols/rendezvous/src/substream_handler.rs
@@ -31,8 +31,7 @@ use instant::Instant;
 use libp2p_core::{InboundUpgrade, OutboundUpgrade, UpgradeInfo};
 use libp2p_swarm::handler::{ConnectionEvent, FullyNegotiatedInbound, FullyNegotiatedOutbound};
 use libp2p_swarm::{
-    ConnectionHandler, ConnectionHandlerEvent, KeepAlive, NegotiatedSubstream, StreamProtocol,
-    SubstreamProtocol,
+    ConnectionHandler, ConnectionHandlerEvent, KeepAlive, Stream, StreamProtocol, SubstreamProtocol,
 };
 use std::collections::{HashMap, VecDeque};
 use std::fmt;
@@ -51,7 +50,7 @@ pub trait SubstreamHandler: Sized {
 
     fn upgrade(open_info: Self::OpenInfo)
         -> SubstreamProtocol<PassthroughProtocol, Self::OpenInfo>;
-    fn new(substream: NegotiatedSubstream, info: Self::OpenInfo) -> Self;
+    fn new(substream: Stream, info: Self::OpenInfo) -> Self;
     fn on_event(self, event: Self::InEvent) -> Self;
     fn advance(self, cx: &mut Context<'_>) -> Result<Next<Self, Self::OutEvent>, Self::Error>;
 }
@@ -541,7 +540,7 @@ impl SubstreamHandler for void::Void {
     type Error = void::Void;
     type OpenInfo = ();
 
-    fn new(_: NegotiatedSubstream, _: Self::OpenInfo) -> Self {
+    fn new(_: Stream, _: Self::OpenInfo) -> Self {
         unreachable!("we should never yield a substream")
     }